// frontend/utils/api/types.ts
import { OutputData } from "@editorjs/editorjs";


export type LoginDto = {
    email: string;
    password: string;
}

export type CreateUserDto = {
    fullName: string;
} & LoginDto;

export type ResponseUser = {
    createdAt: string;
    email: string;
    fullName: string;
    id: number;
    commentsCount?: number;
    postsCount?: number;
    token: string;
    updatedAt: string;
};


export type PostItem = {
    title: string;
    body: OutputData['blocks'];
    description: string;
    tags: null | string;
    id: number;
    views: number;
    user: ResponseUser;
    createdAt: string;
    updatedAt: string;
    commentsCount?: number;
    postsCount?: number;
};

export type CommentItem = {
    id: number;
    text: string;
    post: PostItem;
    user: ResponseUser;
    createdAt: string;
    updatedAt: string;
};